Rearrange pip files to match pip folder structure
